# config/swagger.yml
swagger: "2.0"
info:
  description: "This is a sample API for the Library Management System."
  version: "1.0.0"
  title: "Library API"
host: "localhost:5000"
basePath: ""
schemes:
  - "http"
securityDefinitions:
  BearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
paths:
  /users:
    post:
      summary: "Register a new user"
      description: "Creates a new user with the provided details."
      tags:
        - "Users"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User registration data"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        201:
          description: "User created successfully"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid data or user already exists"
  /users/login:
    post:
      summary: "User Login"
      description: "Login a user and return a JWT token."
      tags:
        - "Users"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User login credentials"
          required: true
          schema:
            type: "object"
            properties:
              email:
                type: "string"
              password:
                type: "string"
      responses:
        200:
          description: "Login successful"
          schema:
            type: "object"
            properties:
              access_token:
                type: "string"
        401:
          description: "Invalid credentials"
...
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        example: 1
      username:
        type: "string"
        example: "john_doe"
      email:
        type: "string"
        example: "john.doe@example.com"
      is_admin:
        type: "boolean"
        example: false
...